+++ spec/active_record_properties_spec.rb
@@ -2,10 +2,6 @@
 
 RSpec.describe ActiveRecordProperties do
   it "has a version number" do
-    expect(ActiveRecordProperties::VERSION).not_to be nil
-  end
-
-  it "does something useful" do
-    expect(false).to eq(true)
+    expect(ActiveRecordProperties::VERSION).not_to be_nil
   end
 end